+++ src/js/Circle.d.ts
@@ -48,6 +48,15 @@ export declare class Circle implements SVGSerializable {
      * @param {number} radius - The radius of the circle.
      */
     constructor(center: Vertex, radius: number);
+    /**
+     * Check if the given circle is fully contained inside this circle.
+     *
+     * @method containsCircle
+     * @param {Circle} circle - The circle to check if it is contained in this circle.
+     * @instance
+     * @memberof Circle
+     * @return {boolean} `true` if any only if the given circle is completely inside this circle.
+     */
     containsCircle(circle: Circle): boolean;
     /**
      * Calculate the distance from this circle to the given line.