+++ src/js/Circle.js
@@ -41,6 +41,15 @@ var Circle = /** @class */ (function () {
         this.radius = radius;
     }
     ;
+    /**
+     * Check if the given circle is fully contained inside this circle.
+     *
+     * @method containsCircle
+     * @param {Circle} circle - The circle to check if it is contained in this circle.
+     * @instance
+     * @memberof Circle
+     * @return {boolean} `true` if any only if the given circle is completely inside this circle.
+     */
     Circle.prototype.containsCircle = function (circle) {
         return this.center.distance(circle.center) + circle.radius < this.radius;
     };