+++ src/js/interfaces.d.ts
@@ -4,6 +4,7 @@ import { Triangle } from "./Triangle";
 import { PBImage } from "./PBImage";
 import { VEllipse } from "./VEllipse";
 import { Circle } from "./Circle";
+import { CircleSector } from "./CircleSector";
 import { Polygon } from "./Polygon";
 import { BezierPath } from "./BezierPath";
 import { Line } from "./Line";
@@ -32,7 +33,7 @@ export interface IBounds {
     min: XYCoords;
     max: XYCoords;
 }
-export declare type Drawable = Vertex | Vector | Triangle | Circle | PBImage | VEllipse | Polygon | BezierPath | Line;
+export declare type Drawable = Vertex | Vector | Triangle | Circle | CircleSector | PBImage | VEllipse | Polygon | BezierPath | Line;
 export interface Config {
     canvas: HTMLCanvasElement | string;
     fullSize?: boolean;
@@ -93,6 +94,7 @@ export interface DrawConfig {
     triangle: DrawSettings;
     ellipse: DrawSettings;
     circle: DrawSettings;
+    circleSector: DrawSettings;
     vertex: DrawSettings;
     selectedVertex: DrawSettings;
     line: DrawSettings;
@@ -117,6 +119,11 @@ export interface SVGSerializable {
         className?: string;
     }) => string;
 }
+/**
+ * A type for SVG &lt;path d="..." /> params.
+ * Example: [ 'A':string, radiusx:number, radiusy:number, rotation:number, largeArcFlag=1|0, sweepFlag=1|0, endx:number, endy:number ]
+ */
+export declare type SVGPathParams = Array<string | number>;
 export interface IHooks {
     saveFile: (pb: PlotBoilerplate) => void;
 }